Kuga

+++ app/src/main/java/com/group6/medicaltest/entity/Nurse.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 )
 data class Nurse(
     @PrimaryKey(autoGenerate = false)
-    val nurseId: Long,
+    val nurseId: Int,
     val firstName: String,
     val lastName: String,
     val department: String,
